Add improved posletter commonality finder util
- This one should be able to find more than one set of commonalities.
Annoyingly it can only find them in pairs of words - and i just can't figure out how to make it group them.
I'd basically want patio, tapir and ratio be grouped - and patio and cow to be grouped.
I just can't think of an approach on how to do that.